requirements.txt
fastmcp
authlib
requests
flask
python-dotenv